Optimize SCSS padding and refactor phone number processing
The styles for the .wp-block-columns class has been simplified by using a single padding attribute. In addition, the way phone numbers are processed in class-contact-infos.php has been streamlined to be more efficient and readable, eliminating unnecessary lines and repetition. This includes both the get_phoneLink and the get_icon methods.